output escaped html in getHTML
