+++ blots/text.js
@@ -2,4 +2,18 @@ import { TextBlot } from 'parchment';
 
 class Text extends TextBlot {}
 
-export default Text;
+function escapeText(text) {
+  return text.replace(/[&<>"']/g, s => {
+    // https://lodash.com/docs#escape
+    const entityMap = {
+      '&': '&amp;',
+      '<': '&lt;',
+      '>': '&gt;',
+      '"': '&quot;',
+      "'": '&#39;',
+    };
+    return entityMap[s];
+  });
+}
+
+export { Text as default, escapeText };
